Mentor comments: 2. CustomUserDetailsService не нужен, пусть UserServiceImpl имплементит одновременно UserService + UserDetailsService 4. пароли  шифруются в сервисе. 6. функциональность StringToRoleConverter влить в RoleService 7. UserDaoImp#getUserByName рефакторинг: кастинга быть не должно, в запросе можно и без алиаса, query не сохраняй в переменную

+++ src/main/java/ru/kata/spring/boot_security/demo/dao/UserDao.java
@@ -7,13 +7,13 @@ import java.util.List;
 public interface UserDao {
     public List<User> getAllUsers();
 
-    public User getUserById(int id);
+    public User getUserById(long id);
 
     public void save(User user);
 
     public void update(User user);
 
-    public void delete(int id);
+    public void delete(long id);
 
     User getUserByName(String username);
 }
