+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserService.java
@@ -9,12 +9,11 @@ public interface UserService {
 
     public User getUserByName(String name);
 
-    public User getUserById(int id);
+    public User getUserById(long id);
 
     public void save(User user);
 
     public void update(User user);
 
-    public void delete(int id);
-
+    public void delete(long id);
 }
